backend/pyproject.toml
[project]
name = "analytical-agent"
version = "0.1.0"
description = "Full-stack analytical platform for MLE, data scientists, and quants"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "duckdb>=1.2.0",
    "polars>=1.20.0",
    "pandas>=2.2.0",
    "altair>=5.5",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "faker>=33.0.0",
    "scipy>=1.14.0",
    "statsmodels>=0.14.0",
    "numpy>=1.26.0",
    "ruptures>=1.1.9",
    "anthropic>=0.40.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "markdown-it-py>=3.0",
    "weasyprint>=62.0",
    "apscheduler>=3.10.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]
second-brain = [
    "second-brain @ file:///Users/jay/Developer/second-brain",
]

[tool.pytest.ini_options]
testpaths = ["tests", "app/skills", "scripts"]
asyncio_mode = "auto"
pythonpath = [".", ".."]
addopts = "--import-mode=importlib"
markers = [
    "slow: marks tests as slow (multi-turn LLM evals — deselect with -m 'not slow')",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.ruff.lint.per-file-ignores]
# Long prompt strings and docstrings with shell commands that cannot be wrapped.
"app/harness/loop.py" = ["E501"]
"scripts/run_eval.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
# type-arg: requires every dict/list to carry generic args — pedantic, not bug-catching.
# unused-ignore: churn-inducing during dependency upgrades.
# no-untyped-call: fires when typed code calls helpers missing annotations.
# no-any-return: noisy when wrapping third-party libs that return Any.
disable_error_code = ["type-arg", "unused-ignore", "no-untyped-call", "no-any-return"]

# Tests: prioritise readability over type strictness. Fixtures and mocks routinely
# pass untyped structures; enforcing strict typing here produces churn without
# catching real bugs. Runtime behaviour is guarded by pytest itself.
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "app.*.tests.*",
    "app.skills.*.tests.*",
    "app.integrity.plugins.*.tests.*",
    "app.skills.*.*.tests.*",
]
ignore_errors = true

# Altair chart builders wrap a third-party API whose Literal-narrowed
# string arguments and LayerChart/Chart return-type unions are awkward to
# satisfy without verbose casts. Behaviour is exercised by skill evals.
[[tool.mypy.overrides]]
module = ["app.skills.charting.altair_charts.*"]
disable_error_code = ["arg-type", "return-value"]

# Stat fixtures are interactive notebooks-as-code — annotations would obscure intent.
[[tool.mypy.overrides]]
module = ["app.skills._stat_fixtures.*"]
disallow_untyped_defs = false

# Third-party libraries without type stubs.
[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "scipy.*",
    "ruptures.*",
    "apscheduler.*",
    "second_brain.*",
    "statsmodels.*",
    "weasyprint.*",
    "altair.*",
    "markdown_it.*",
    "faker.*",
    "prometheus_fastapi_instrumentator.*",
    "duckdb.*",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.15.10",
    "types-pyyaml>=6.0.12.20260408",
    "vulture>=2.16",
]
